downgrade notify component to dumb
